Clean up address resolution, add more insecurities ;p
Switch to using just the old fashioned getaddrinfo(3) to resolve
addresses, however this does push the use of char* for the ports...but
meh?

Now it works with both ip addresses and hostnames, Huzzah?

On that note, before you couldn't use an ip address for connecting,
but now you can. This also adds the "noverifyname" insecurity (already
had "noverifycert") so self-signed certs don't have any verification
but their expirations.