Resolve some incorrect logic covering off-by-1 error
So my odd, arbitrary appending of NUL's was covering up some bad
math. This makes the init_frame() function return the current length
of the frame in bytes, which makes more sense. The odd addition of
null bytes can now be dropped, but the dumb_frame() accounting needed
some tweaking as well.

In the spirit of off-by-1 and null-terminator issues, fix the
client_test.c code to properly deal with the fact we aren't sending
the NULL terminator in the payload (since it makes nodejs
pissy). Since our payload is arbitrary data and not a string, it makes
no sense to be appending them like the code was anyways.