Remove some BSD'isms (#1), fix up Makefile
Remove use of err(3), replacing it with writing to stderr and
exit(3)-ing.

Fix up the Makefile logic so building certs is a critical step now for
testing. Also run both non-TLS and TLS tests.

As a consequence of removing err(3), use assert(3) now in the test
since most of the functions exit(3) on real errors (for now).

Close #1