"friendship ended with nodejs, now go is my friend"
This imports and vendors the Go Gorilla Websocket module.

It's got the BSD-2-Clause license, so I'm cool with pulling it into this
tree to have a dedicated websocket test bed. This does mean I need to
add TLS support and a Makefile, but I think this will give me a nice way
to automate testing of the dws.c code.

No more expecting someone to run `npm i` or `cargo build`!