Bye official arc4random usage :'(
Too much of a hassle to deal with ifdefs. At least on OpenBSD I know
it's being used under the covers. The dumb_key generation isn't really
that important, so I'm not worried about modulo-bias issues.

Also updates the docs to reflect new lack of dependency on that build
time flag.