Time sucks as a seed, use /dev/urandom.
Here I go maybe caring a bit too much...

Use initstate(3) to properly initialize use of random(3).

For now, use 256 bytes from /dev/urandom...in the future might need to
provide a means to pass in the state in the event the process
performing the dumb websocket client connection isn't allowed access
to /dev/urandom or the open(2) syscall.

(This is a bigger design issue as my crap() function writes to
stderr...so this isn't quite a good library design yet.)