Add echo mode to go-test so it won't block
Not doing anything fancy with go routines, but it's easily possible to
deadlock the client side using my original go server by writing
responses that never get consumed. Simply not doing so resolves it, so
this adds an 'echo' flag (-e) that will enable echoing the client's
message back to it...otherwise it stays mute.